requirements.txt
pandas
numpy
backtesting
python-dotenv
python-binance
backtesting
TA-lib